WIP - Allow up to 4 items - starting to limp along
I have made a quick and dirty pass through and moved a lot of the data items into an array of them.  Currently set as max of 4.

Currently only testing with already previously paired devices. but now showing data, when I connect up Keyboard (hit some key)...

Then starts up mouse, press a button on it.  Now the code is giving the keyboard events to the keyboard object and mouse  events to the mouse controller...