Make Keyboard object always rely on HIDParser object(s) and no longer top level object
I always disliked the code in hid.cpp that checked for keyboard boot and then did not claim and then have keyboard claim it, as some keyboards may have boot marked, but don't fully use it by default in boot mode.

So with this version I made the code such that HID can and does claim it.
Note: most keyboards will probably claim two HIDParsers.  One for the boot interface and one for the Multimedia and system pages.

The code detects when new data comes in if the data is 8 bytes long and if so processes it using the boot code, else it lets it go through and be processed by HID code.

Note: I first tried to let the HID parser actually parse the data, but it had issues, so did as I mentioned