Create MouseKeyboardForward.ino
Added my extended test program to allow both Mouse and or Keyboard to be forwarded.

It is setup that if you are running USB Type== Serial
it will simply output information about the mouse or keyboard events.

If the usb type include keyboard, it will forward to keyboard.
ditto for mouse -> mouse