Warning: Rearranging code - lots probably busted
I am in the process of changing the Bluetooth connections code.

In particular, I have changed the embedded array structure to be a top level class.  Note: I still have two of them I add by default, can add more of the objects if you need them in your sketch.  I am moving functionality from the main BT object to these.  I am trying to have these objects be the one that talk to the BTHID objects, as they need the connection information, and not rely on hopefully the last thing done was for their connection.

Only part way there.  It is also having some new methods for processing HID on the BTHID object.  SO far only Keyboard has implemented any of this.

More to come!