RX2_data - two different ways to need to continue to read buffer
One is where the HCI packet length is such that it wont fit in 64 bytes...
We were not handling this.  For RX_read we did read in the continue data.

before now we through away the data that did not fit ifirst packet.

Second way is HCI length is < LCAP length in which case we expect another real message, with the 0x10 bit set in 2nd byte and then we need to combine the data.

I added this earlier, but it interfered with the first case, which again before we just tossed the data.  Now I combine it like we do in the RX case.